rename: grouping configuration files
