+++ src/main/java/com/capstone/popup/global/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.capstone.popup.global;
+package com.capstone.popup.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
